+++ tests/test_project.py
@@ -46,12 +46,14 @@ def test_task():
     from tasksched import Task
     with pytest.raises(TypeError):
         Task()
-    task = Task('id', 'the title', 30, 3)
+    task = Task('id', 'the title', 30, 100, 3)
     assert task.task_id == 'id'
     assert task.title == 'the title'
     assert task.duration == 30
+    assert task.priority == 100
     assert task.max_resources == 3
-    assert str(task) == 'Task id - the title: 30d (max res: 3)'
+    assert str(task) == ('Task id - the title: 30d, priority: 100, '
+                         'max resources: 3')
 
 
 def test_project():
@@ -94,26 +96,46 @@ def test_project():
     assert project.tasks[0].task_id == 'task1'
     assert project.tasks[0].title == 'The first task'
     assert project.tasks[0].duration == 2
+    assert project.tasks[0].priority == 0
     assert project.tasks[0].max_resources == 2
     assert project.tasks[1].task_id == 'task2'
     assert project.tasks[1].title == 'The second task'
     assert project.tasks[1].duration == 5
+    assert project.tasks[1].priority == 0
     assert project.tasks[1].max_resources == 2
     assert project.tasks[2].task_id == 'task3'
     assert project.tasks[2].title == 'The third task'
     assert project.tasks[2].duration == 10
+    assert project.tasks[2].priority == 0
     assert project.tasks[2].max_resources == 2
-    tasks_by_duration = project.tasks_by_duration()
-    assert tasks_by_duration[0].task_id == 'task3'
-    assert tasks_by_duration[1].task_id == 'task2'
-    assert tasks_by_duration[2].task_id == 'task1'
     assert str(project) == """\
 Project: The name
   Resources:
     Resource dev1 - Developer 1
     Resource dev2 - Developer 2
   Tasks:
-    Task task1 - The first task: 2d (max res: 2)
-    Task task2 - The second task: 5d (max res: 2)
-    Task task3 - The third task: 10d (max res: 2)
+    Task task1 - The first task: 2d, priority: 0, max resources: 2
+    Task task2 - The second task: 5d, priority: 0, max resources: 2
+    Task task3 - The third task: 10d, priority: 0, max resources: 2
 """
+
+
+def test_project_sort_tasks():
+    """Test sort of tasks in a project."""
+    from tasksched import Project
+    project = Project(get_json_file('project_complete.json'))
+    sorted_tasks = project.sorted_tasks(['duration'])
+    assert sorted_tasks[0].task_id == 'task1'
+    assert sorted_tasks[1].task_id == 'task2'
+    assert sorted_tasks[2].task_id == 'task3'
+    project = Project(get_json_file('project_complete.json'))
+    sorted_tasks = project.sorted_tasks(['duration'], reverse=True)
+    assert sorted_tasks[0].task_id == 'task3'
+    assert sorted_tasks[1].task_id == 'task2'
+    assert sorted_tasks[2].task_id == 'task1'
+    project = Project(get_json_file('project_complete2.json'))
+    sorted_tasks = project.sorted_tasks(['priority', 'duration'],
+                                        reverse=True)
+    assert sorted_tasks[0].task_id == 'task2'
+    assert sorted_tasks[1].task_id == 'task3'
+    assert sorted_tasks[2].task_id == 'task1'